Move OutputIdProof to block types
