+++ sdk/src/wallet/types/mod.rs
@@ -19,9 +19,9 @@ pub use self::{
 use crate::{
     client::secret::types::InputSigningData,
     types::{
-        api::core::{OutputIdProof, OutputWithMetadataResponse},
+        api::core::OutputWithMetadataResponse,
         block::{
-            output::{Output, OutputId, OutputMetadata},
+            output::{Output, OutputId, OutputIdProof, OutputMetadata},
             payload::signed_transaction::{dto::SignedTransactionPayloadDto, SignedTransactionPayload, TransactionId},
             protocol::{CommittableAgeRange, ProtocolParameters},
             slot::SlotIndex,
